Move handleLayoutTransition
